feat: create Sidebar component
